Create LOAD_OPPR_ALL_DTL_GNRL_12192014
Creates the table structure and loads the records from an S3 csv file downloaded from CMS

File #1 - OPPR_ALL_DTL_GNRL_12192014.csv: 
This file contains the data set for identified General Payments for the 2013 program year, including the additional payment records added to the publication during the data refresh in December 2014. General Payments are defined as payments or other transfers of value not made in connection with a research agreement or research protocol.